requirements_visual_rnn.txt
numpy
scipy
matplotlib
scikit-learn
torch
tensorboard
shapely
ratinabox
umap-learn
networkx
pandas
ripser
PyYAML
